Introduce a mockable GisSdkClient for tests.
